All changes  are done
Manual test checklist
Wholesale grid — Open Wholesale, add items; confirm there is no Revenue/Cost/Profit column (columns end at Total → Del).
Wholesale print — Print a wholesale bill; header/title reads "Bill / Invoice" (no "Wholesale"), and layout/columns/discount/date match a normal Sale bill exactly.
Demand Order grid — Open from Dashboard low-stock; confirm columns are exactly ✓ | Serial No | Item | Quantity | Price, all rows checked by default, Quantity & Price editable.
Demand Order print — Uncheck a couple of rows, Print; only checked rows appear, renumbered 1..n, columns S.No | Item | Qty | Price, Price column blank, no checkbox, no Total line.
Sales search — manual — In normal Sale, type "g"; confirm nothing is added until you press Enter or click a dropdown item. Verify Escape after opening the dropdown adds nothing.
Sales search — scanner — Scan a barcode into the Barcode field; item drops straight into the grid as before.
Product price decimals — In Product Management, type 125.5 / 100.25 into Cost/Retail/Wholesale price; confirm the decimal point is accepted and saved (was previously blocked).
Decimal quantity — After applying the migration, on both Sale and Wholesale enter qty 1.3 / 1.5 in the grid; confirm it's accepted, line total = price × decimal qty, and it saves/prints correctly.
Recommend running /verify (drive the Sale + Product Management flows) before committing, since these are UI-behavior changes that unit tests won't fully exercise.

+++ POSApp.Core/Interfaces/ISaleRepository.cs
@@ -26,7 +26,7 @@ namespace POSApp.Core.Interfaces
     public sealed class SalesByCategoryDto
     {
         public string CategoryName { get; set; } = string.Empty;
-        public int TotalQuantity { get; set; }
+        public decimal TotalQuantity { get; set; }
         public decimal TotalSales { get; set; }
         public decimal TotalProfit { get; set; }
     }
@@ -35,7 +35,7 @@ namespace POSApp.Core.Interfaces
     {
         public string ProductId { get; set; } = string.Empty;
         public string ProductName { get; set; } = string.Empty;
-        public int TotalQuantity { get; set; }
+        public decimal TotalQuantity { get; set; }
         public decimal TotalSales { get; set; }
         public decimal TotalProfit { get; set; }
     }
