+++ POSApp.UI/ViewModels/SaleReturnViewModel.cs
@@ -167,7 +167,7 @@ namespace POSApp.UI.ViewModels
                 {
                     ProductId = item.ProductId,
                     ProductName = item.ProductName,
-                    OriginalQuantity = item.Quantity,
+                    OriginalQuantity = (int)item.Quantity,
                     ReturnQuantity = 0,
                     UnitPrice = item.UnitPrice,
                     CostPrice = item.CostPrice,
